Replace empty asset files during rip
Some empty asset files get created if you try to build the ROM prior to ripping assets. Empty asset files will now be replaced if you subsequently rip assets so that you don't have to run `make distclean` first. This is a pretty minor fix but could be helpful is someone accidentally runs the make commands in the wrong order.